Add tests

+++ build.ps1
@@ -1,12 +1,23 @@
 #requires -Version 6.0
 
+[CmdletBinding(DefaultParameterSetName = 'Build')]
 param(
+    [Parameter(ParameterSetName = 'Build')]
     [ValidateSet('Debug', 'Release')]
     [string]
     $Configuration = 'Debug',
 
+    [Parameter(ParameterSetName = 'Build')]
     [switch]
-    $Clean
+    $Clean,
+
+    [Parameter(ParameterSetName = 'Test')]
+    [switch]
+    $Build,
+
+    [Parameter(ParameterSetName = 'Test')]
+    [switch]
+    $Test
 )
 
 $ErrorActionPreference = 'Stop'
@@ -44,42 +55,52 @@ function Exec([scriptblock]$sb, [switch]$IgnoreExitcode)
     }
 }
 
-try
+if ($PSCmdlet.ParameterSetName -eq 'Build' -or $Build)
 {
-    $null = Get-Command dotnet -ErrorAction Stop
-}
-catch
-{
-    throw 'Unable to find dotnet executable'
-}
-
-if ($Clean)
-{
-    foreach ($path in $script:OutDir,"${script:SrcDir}/bin","${script:SrcDir}/obj")
+    try
     {
-        if (Test-Path -Path $path)
+        $null = Get-Command dotnet -ErrorAction Stop
+    }
+    catch
+    {
+        throw 'Unable to find dotnet executable'
+    }
+
+    if ($Clean)
+    {
+        foreach ($path in $script:OutDir,"${script:SrcDir}/bin","${script:SrcDir}/obj")
         {
-            Remove-Item -Force -Recurse -Path $path -ErrorAction Stop
+            if (Test-Path -Path $path)
+            {
+                Remove-Item -Force -Recurse -Path $path -ErrorAction Stop
+            }
         }
     }
+
+    Push-Location $script:SrcDir
+    try
+    {
+        Exec { dotnet build }
+    }
+    finally
+    {
+        Pop-Location
+    }
+
+    New-Item -ItemType Directory -Path $script:OutModuleDir -ErrorAction SilentlyContinue
+
+    foreach ($artifactEntry in $script:Artifacts.GetEnumerator())
+    {
+        Copy-Item -Path $artifactEntry.Key -Destination (Join-Path $script:OutModuleDir $artifactEntry.Value) -ErrorAction Stop
+    }
+
+    # We need the zsh completer script to drive zsh completions
+    Invoke-WebRequest -Uri 'https://raw.githubusercontent.com/Valodim/zsh-capture-completion/master/capture.zsh' -OutFile $script:ZshCompleterScriptLocation
 }
 
-Push-Location $script:SrcDir
-try
+if ($Test)
 {
-    Exec { dotnet build }
+    $pwsh = (Get-Process -Id $PID).Path
+    $testPath = "$PSScriptRoot/tests"
+    & $pwsh -c "Invoke-Pester '$testPath'"
 }
-finally
-{
-    Pop-Location
-}
-
-New-Item -ItemType Directory -Path $script:OutModuleDir -ErrorAction SilentlyContinue
-
-foreach ($artifactEntry in $script:Artifacts.GetEnumerator())
-{
-    Copy-Item -Path $artifactEntry.Key -Destination (Join-Path $script:OutModuleDir $artifactEntry.Value) -ErrorAction Stop
-}
-
-# We need the zsh completer script to drive zsh completions
-Invoke-WebRequest -Uri 'https://raw.githubusercontent.com/Valodim/zsh-capture-completion/master/capture.zsh' -OutFile $script:ZshCompleterScriptLocation
